getParameters is a proper function instead of monkeypatched onto collections
