+++ packages/lesswrong/lib/vulcan-lib/collections.ts
@@ -1,9 +1,6 @@
 import { MongoCollection } from '../mongoCollection';
 import PgCollection from '../sql/PgCollection';
 import SwitchingCollection from '../SwitchingCollection';
-import * as _ from 'underscore';
-import merge from 'lodash/merge';
-import { DatabasePublicSetting } from '../publicSettings';
 import { getDefaultFragmentText, registerFragment } from './fragments';
 import { registerCollection } from './getCollection';
 import { addGraphQLCollection } from './graphql';
@@ -11,9 +8,6 @@ import { camelCaseify } from './utils';
 import { pluralize } from './pluralize';
 export * from './getCollection';
 import { loggerConstructor } from '../utils/logging'
-
-// 'Maximum documents per request'
-const maxDocumentsPerRequestSetting = new DatabasePublicSetting<number>('maxDocumentsPerRequest', 5000)
 
 // When used in a view, set the query so that it returns rows where a field is
 // null or is missing. Equivalent to a search with mongo's `field:null`, except
@@ -109,90 +103,6 @@ export const createCollection = <
   const defaultFragment = getDefaultFragmentText(collection, schema);
   if (defaultFragment) registerFragment(defaultFragment);
 
-  // ------------------------------------- Parameters -------------------------------- //
-
-  collection.getParameters = ((terms: ViewTermsByCollectionName[N] = {}, apolloClient?: any, context?: ResolverContext): MergedViewQueryAndOptions<N,T> => {
-    const logger = loggerConstructor(`views-${collectionName.toLowerCase()}`)
-    logger('getParameters(), terms:', terms);
-
-    let parameters: any = {
-      selector: {},
-      options: {},
-    };
-
-    if (collection.defaultView) {
-      parameters = merge(
-        parameters,
-        collection.defaultView(terms, apolloClient, context)
-      );
-      logger('getParameters(), parameters after defaultView:', parameters)
-    }
-
-    // handle view option
-    if (terms.view && collection.views[terms.view]) {
-      const viewFn = collection.views[terms.view];
-      const view = viewFn(terms, apolloClient, context);
-      let mergedParameters = merge(parameters, view);
-
-      if (
-        mergedParameters.options &&
-        mergedParameters.options.sort &&
-        view.options &&
-        view.options.sort
-      ) {
-        // If both the default view and the selected view have sort options,
-        // don't merge them together; take the selected view's sort. (Otherwise
-        // they merge in the wrong order, so that the default-view's sort takes
-        // precedence over the selected view's sort.)
-        mergedParameters.options.sort = view.options.sort;
-      }
-      parameters = mergedParameters;
-      logger('getParameters(), parameters after defaultView and view:', parameters)
-    }
-
-    // sort using terms.orderBy (overwrite defaultView's sort)
-    if (terms.orderBy && !_.isEmpty(terms.orderBy)) {
-      parameters.options.sort = terms.orderBy;
-    }
-
-    // if there is no sort, default to sorting by createdAt descending
-    if (!parameters.options.sort) {
-      parameters.options.sort = { createdAt: -1 } as any;
-    }
-
-    // extend sort to sort posts by _id to break ties, unless there's already an id sort
-    // NOTE: always do this last to avoid overriding another sort
-    if (!(parameters.options.sort && typeof parameters.options.sort._id !== undefined)) {
-      parameters = merge(parameters, { options: { sort: { _id: -1 } } });
-    }
-
-    // remove any null fields (setting a field to null means it should be deleted)
-    _.keys(parameters.selector).forEach(key => {
-      if (_.isEqual(parameters.selector[key], viewFieldNullOrMissing)) {
-        parameters.selector[key] = null;
-      } else if (_.isEqual(parameters.selector[key], viewFieldAllowAny)) {
-        delete parameters.selector[key];
-      } else if (parameters.selector[key] === null || parameters.selector[key] === undefined) {
-        //console.log(`Warning: Null key ${key} in query of collection ${collectionName} with view ${terms.view}.`);
-        delete parameters.selector[key];
-      }
-    });
-    if (parameters.options.sort) {
-      _.keys(parameters.options.sort).forEach(key => {
-        if (parameters.options.sort[key] === null) {
-          delete parameters.options.sort[key];
-        }
-      });
-    }
-
-    // limit number of items to 1000 by default
-    const maxDocuments = maxDocumentsPerRequestSetting.get();
-    const limit = terms.limit || parameters.options.limit;
-    parameters.options.limit = !limit || limit < 1 || limit > maxDocuments ? maxDocuments : limit;
-
-    logger('getParameters(), final parameters:', parameters);
-    return parameters;
-  }) as any;
 
   registerCollection(collection);
